requirements.txt
mysql-connector-python
psycopg2